+++ src/engineui/windows/diagnostics_window.h
@@ -9,7 +9,7 @@ class DiagnosticsWindow : public EngineWindow
 public:
 	void prepare();
 private:
-	static std::deque<int> fpsCache;
+	static std::deque<float> fpsCache;
 	static float fpsUpdateTimer;
 };
